# Conflicts:
#	include/etl/iterator.h
#	include/etl/platform.h
#	test/vs2019/etl.vcxproj
#	test/vs2019/etl.vcxproj.filters

+++ test/test_string_stream_wchar_t.cpp
@@ -68,7 +68,7 @@ namespace etl
   {
     for (auto c : str)
     {
-      os << c;
+      os << uint16_t(c);
     }
 
     return os;
